requirements.txt
langgraph>=0.2
langchain-core>=0.3
openai>=1.0
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
